backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-openai

uuid
langchain-community
zep-cloud
rich
